+++ changelog.md
@@ -7,6 +7,8 @@
 
 #### v11.03
 
+- Added command line option to edit script config
+- Moved configs from kamehouse.cfg to script specific config files
 - Updated shell script config paths to use .kamehouse folder
 - Fixed bugs